telegram_bot/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.7.0
async-timeout==4.0.2
attrs==23.1.0
black==23.7.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
dnspython==2.4.0
email-validator==2.0.0.post2
exceptiongroup==1.1.1
fastapi
frozenlist==1.4.0
gunicorn==20.1.0
h11==0.14.0
httpcore==0.17.3
httptools==0.5.0
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
mypy-extensions==1.0.0
orjson==3.9.2
packaging==23.1
pathspec==0.11.1
platformdirs==3.9.1
pluggy==1.2.0
pydantic==1.10.9
pyngrok==6.0.0
pytest==7.4.0
python-dotenv==1.0.0
python-multipart==0.0.6
python-telegram-bot==20.4
PyYAML==6.0
requests==2.31.0
sniffio==1.3.0
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.6.3
ujson==5.8.0
urllib3==2.0.3
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
xmltodict==0.13.0
yarl==1.9.2
fastapi-cors
feedparser